chore(dal): split func backends into modules
Split existing func backends into their own modules, like validations.
This is done for usability and was encountered when adding new prop
types.
