+++ lib/dal/src/func/backend.rs
@@ -1,19 +1,15 @@
-use std::collections::HashMap;
-
 use serde::{Deserialize, Serialize};
 use strum_macros::{AsRefStr, Display, EnumIter, EnumString};
 use telemetry::prelude::*;
 use thiserror::Error;
-use tokio::sync::mpsc;
-
-use veritech::{
-    Client, FunctionResult, OutputStream, QualificationCheckComponent, QualificationCheckRequest,
-    ResolverFunctionRequest, ResourceSyncComponent, ResourceSyncRequest,
-};
 
 use crate::{edit_field::ToSelectWidget, label_list::ToLabelList};
 
 pub mod integer;
+pub mod js_qualification;
+pub mod js_resource;
+pub mod js_string;
+pub mod string;
 pub mod validation;
 
 #[derive(Error, Debug)]
@@ -96,273 +92,3 @@ pub enum FuncBackendResponseType {
 
 impl ToLabelList for FuncBackendKind {}
 impl ToSelectWidget for FuncBackendKind {}
-
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct FuncBackendStringArgs {
-    pub value: String,
-}
-
-impl FuncBackendStringArgs {
-    pub fn new(value: String) -> Self {
-        Self { value }
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct FuncBackendString {
-    args: FuncBackendStringArgs,
-}
-
-impl FuncBackendString {
-    pub fn new(args: FuncBackendStringArgs) -> Self {
-        Self { args }
-    }
-
-    #[instrument(
-        name = "funcbackendstring.execute",
-        skip_all,
-        level = "debug",
-        fields(
-            otel.kind = %SpanKind::Client,
-            otel.status_code = Empty,
-            otel.status_message = Empty,
-            si.func.result = Empty
-        )
-    )]
-    pub async fn execute(self) -> FuncBackendResult<serde_json::Value> {
-        let span = Span::current();
-
-        let value = serde_json::to_value(&self.args.value)?;
-        // You can be damn sure this is a string, really - because
-        // the inner type there is a string. But hey - better safe
-        // than sorry!
-        if !value.is_string() {
-            return Err(span.record_err(FuncBackendError::InvalidStringData(value)));
-        }
-
-        span.record_ok();
-        span.record("si.func.result", &tracing::field::debug(&value));
-        Ok(value)
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct FuncBackendJsStringArgs {
-    pub arguments: HashMap<String, serde_json::Value>,
-}
-
-#[derive(Debug)]
-pub struct FuncBackendJsString {
-    veritech: Client,
-    output_tx: mpsc::Sender<OutputStream>,
-    request: ResolverFunctionRequest,
-}
-
-impl FuncBackendJsString {
-    pub fn new(
-        veritech: Client,
-        output_tx: mpsc::Sender<OutputStream>,
-        handler: impl Into<String>,
-        args: HashMap<String, serde_json::Value>,
-        code_base64: impl Into<String>,
-    ) -> Self {
-        let request = ResolverFunctionRequest {
-            // Once we start tracking the state of these executions, then this id will be useful,
-            // but for now it's passed along and back, and is opaue
-            execution_id: "tomcruise".to_string(),
-            handler: handler.into(),
-            parameters: Some(args),
-            code_base64: code_base64.into(),
-        };
-
-        Self {
-            veritech,
-            output_tx,
-            request,
-        }
-    }
-
-    #[instrument(
-        name = "funcbackendjsstring.execute",
-        skip_all,
-        level = "debug",
-        fields(
-            otel.kind = %SpanKind::Client,
-            otel.status_code = Empty,
-            otel.status_message = Empty,
-            si.func.result = Empty
-        )
-    )]
-    pub async fn execute(self) -> FuncBackendResult<serde_json::Value> {
-        let span = Span::current();
-
-        let result = self
-            .veritech
-            .execute_resolver_function(self.output_tx, &self.request)
-            .await
-            .map_err(|err| span.record_err(err))?;
-        let value = match result {
-            FunctionResult::Success(success) => {
-                if success.unset {
-                    return Err(span.record_err(FuncBackendError::UnexpectedUnset));
-                }
-                if !success.data.is_string() {
-                    return Err(span.record_err(FuncBackendError::InvalidStringData(success.data)));
-                }
-                success.data
-            }
-            FunctionResult::Failure(failure) => {
-                return Err(span.record_err(FuncBackendError::ResultFailure {
-                    kind: failure.error.kind,
-                    message: failure.error.message,
-                }));
-            }
-        };
-
-        span.record_ok();
-        span.record("si.func.result", &tracing::field::debug(&value));
-        Ok(value)
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct FuncBackendJsQualificationArgs {
-    pub component: QualificationCheckComponent,
-}
-
-#[derive(Debug)]
-pub struct FuncBackendJsQualification {
-    veritech: Client,
-    output_tx: mpsc::Sender<OutputStream>,
-    request: QualificationCheckRequest,
-}
-
-impl FuncBackendJsQualification {
-    pub fn new(
-        veritech: Client,
-        output_tx: mpsc::Sender<OutputStream>,
-        handler: impl Into<String>,
-        component: QualificationCheckComponent,
-        code_base64: impl Into<String>,
-    ) -> Self {
-        let request = QualificationCheckRequest {
-            // Once we start tracking the state of these executions, then this id will be useful,
-            // but for now it's passed along and back, and is opaue
-            execution_id: "danielcraig".to_string(),
-            handler: handler.into(),
-            component,
-            code_base64: code_base64.into(),
-        };
-
-        Self {
-            veritech,
-            output_tx,
-            request,
-        }
-    }
-
-    #[instrument(
-        name = "funcbackendjsqualification.execute",
-        skip_all,
-        level = "debug",
-        fields(
-            otel.kind = %SpanKind::Client,
-            otel.status_code = Empty,
-            otel.status_message = Empty,
-            si.func.result = Empty
-        )
-    )]
-    pub async fn execute(self) -> FuncBackendResult<serde_json::Value> {
-        let span = Span::current();
-
-        let result = self
-            .veritech
-            .execute_qualification_check(self.output_tx, &self.request)
-            .await
-            .map_err(|err| span.record_err(err))?;
-        let value = match result {
-            FunctionResult::Success(check_result) => serde_json::to_value(&check_result)?,
-            FunctionResult::Failure(failure) => {
-                return Err(span.record_err(FuncBackendError::ResultFailure {
-                    kind: failure.error.kind,
-                    message: failure.error.message,
-                }));
-            }
-        };
-
-        span.record_ok();
-        span.record("si.func.result", &tracing::field::debug(&value));
-        Ok(value)
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct FuncBackendJsResourceSyncArgs {
-    pub component: ResourceSyncComponent,
-}
-
-#[derive(Debug)]
-pub struct FuncBackendJsResourceSync {
-    veritech: Client,
-    output_tx: mpsc::Sender<OutputStream>,
-    request: ResourceSyncRequest,
-}
-
-impl FuncBackendJsResourceSync {
-    pub fn new(
-        veritech: Client,
-        output_tx: mpsc::Sender<OutputStream>,
-        handler: impl Into<String>,
-        component: ResourceSyncComponent,
-        code_base64: impl Into<String>,
-    ) -> Self {
-        let request = ResourceSyncRequest {
-            // Once we start tracking the state of these executions, then this id will be useful,
-            // but for now it's passed along and back, and is opaue
-            execution_id: "rodrigosantoro".to_string(),
-            handler: handler.into(),
-            component,
-            code_base64: code_base64.into(),
-        };
-
-        Self {
-            veritech,
-            output_tx,
-            request,
-        }
-    }
-
-    #[instrument(
-        name = "funcbackendjsresourcesync.execute",
-        skip_all,
-        level = "debug",
-        fields(
-            otel.kind = %SpanKind::Client,
-            otel.status_code = Empty,
-            otel.status_message = Empty,
-            si.func.result = Empty
-        )
-    )]
-    pub async fn execute(self) -> FuncBackendResult<serde_json::Value> {
-        let span = Span::current();
-
-        let result = self
-            .veritech
-            .execute_resource_sync(self.output_tx, &self.request)
-            .await
-            .map_err(|err| span.record_err(err))?;
-        let value = match result {
-            FunctionResult::Success(result) => serde_json::to_value(&result)?,
-            FunctionResult::Failure(failure) => {
-                return Err(span.record_err(FuncBackendError::ResultFailure {
-                    kind: failure.error.kind,
-                    message: failure.error.message,
-                }));
-            }
-        };
-
-        span.record_ok();
-        span.record("si.func.result", &tracing::field::debug(&value));
-        Ok(value)
-    }
-}
