+++ lib/dal/src/schema/builtins.rs
@@ -1,6 +1,6 @@
+use crate::func::backend::js_qualification::FuncBackendJsQualificationArgs;
+use crate::func::backend::js_resource::FuncBackendJsResourceSyncArgs;
 use crate::func::backend::validation::FuncBackendValidateStringValueArgs;
-use crate::func::backend::FuncBackendJsQualificationArgs;
-use crate::func::backend::FuncBackendJsResourceSyncArgs;
 use crate::qualification_prototype::QualificationPrototypeContext;
 use crate::resource_prototype::ResourcePrototypeContext;
 use crate::schema::{SchemaResult, SchemaVariant, UiMenu};
